requirements.txt
fastapi
uvicorn
pydantic
pdfplumber
pytesseract
Pillow
pdf2image
python-docx
sentence-transformers
faiss-cpu
numpy
google-generativeai
python-multipart
